\name{checkData}
\alias{checkData}

\title{Check data consistency}
\description{
Check data consistency among traits.
}
\usage{
checkData(phenoDTfile, crop)
}
\arguments{
\item{myObject}{The phenotypic data frame generated by \code{checkNames}.}
\item{crop}{A character string to identify the crop (\code{potato} or \code{sweetpotato}).}
}
\details{
The data frame must use the labels (lower or upper case) listed it \code{?st4gi::check.names.pt} and
\code{?st4gi::check.names.sp}.
}
\value{
It returns all rows with some kind of inconsistency among traits in data.frame format.
}

\examples{
####=========================================####
data(DT_example)

tmp <- checkData(phenoDTfile = result$data$pheno, crop = 'potato')

}

