\name{checkData}
\alias{checkData}

\title{Check data consistency}
\description{
Check data consistency among traits.
}
\usage{
checkData(crop, phenoDTfile, f, print.text)
}
\arguments{
\item{crop}{A character string to identify the crop (\code{potato} or \code{sweetpotato}).}
\item{myObject}{The phenotypic data frame generated by \code{checkNames}.}
\item{f}{Factor for extreme values detection, set to \code{f = 0} to do not run.} 
\item{print.text}{Logical, if \code{TRUE} the output is printed on screen.} 
}
\details{
The data frame must use the labels (lower or upper case) listed it \code{?st4gi::check.names.pt} and
\code{?st4gi::check.names.sp}.
}
\value{
It returns:
\itemize{
\item \code{$Inconsist.List}, a \code{data.frame} with a list of all the
rows with some kind of inconsistency.
\item \code{$Inconsist.Matrix}, a \code{data.frame} with the positions
in the fieldbook data frame where inconsistencies occur. These are coded
with: (1) for inconsistencies among traits and (2) for out of range values.
}
}

\examples{
####=========================================####
data(DT_example)

tmp <- checkData(crop = 'potato', phenoDTfile = result$data$pheno)

}

